pyproject.toml
[project]
name = "failsafe-bot"
version = "v1.0.7"
description = "Telegram and Discord chat bot"
requires-python = "==3.12.9"
dependencies = [
    "discord-py>=2.4.0",
    "numpy>=2.1.1",
    "psutil>=6.0.0",
    "pynacl>=1.5.0",
    "python-telegram-bot>=21.5",
    "toml>=0.10.2",
    "tqdm>=4.66.5",
    "unidecode>=1.3.8",
    "yt-dlp>=2024.8.6",
    "httpx==0.27.2",
    "openai==1.55.3",
    "requests>=2.32.3",
    "strsimpy>=0.2.1",
    "jsonlines>=4.0.0",
    "tiktoken>=0.9.0",
]
